feat(website): use shared worker for event api
