+++ apps/website/package.json
@@ -23,6 +23,7 @@
 		"@sveltejs/adapter-static": "^3.0.6",
 		"@sveltejs/kit": "^2.8.0",
 		"@sveltejs/vite-plugin-svelte": "^4.0.0",
+		"@types/sharedworker": "^0.0.132",
 		"@urql/exchange-auth": "^2.2.0",
 		"@urql/svelte": "^4.2.2",
 		"eslint": "^9.14.0",
